sudo: add commands allowed to be called by ssh_group
Problem: getDebugInformation asks for a password if it is called by non-root users which is now the default (ssh_wan or ssh_lan).

Solution: add the commands used by getDebugInformation to the list of commands allowed to be called without a password by users part of the ssh_group.

Issue: HOP-9644 [SSH] getDebugInformation asks for a password
GitOrigin-RevId: e940b6262649f4163694c5d771807b3426ceb463